requirements.txt
     bddl==1.0.1
     click==8.1.7
     cloudpickle==1.6.0
     cmake==3.27.2
     deepspeed==0.11.1
     diffusers==0.20.1
     easydict==1.10
     egl-probe==1.0.2
     einops==0.6.1
     environs==9.5.0
     flow-vis==0.1
     future==0.18.2
     gdown==4.7.1
     glfw==2.6.2
     gym==0.22.0
     gymnasium==0.29.1
     h5py==3.9.0
     huggingface-hub==0.16.4
     hydra-core==1.3.2
     imageio==2.31.2
     imageio-ffmpeg==0.4.9
     ipdb==0.13.13
     ipykernel==6.27.1
     ipython==8.12.2
     json5==0.9.14
     jupyterlab==4.0.9
     keras==2.13.1
     lightning==2.1.0
     lightning-utilities==0.9.0
     line-profiler==4.1.1
     lit==16.0.6
     llvmlite==0.40.1
     matplotlib==3.7.2
     matplotlib-inline==0.1.6
     mediapy==1.2.0
     moviepy==1.0.3
     mujoco==3.2.3
     natsort==8.4.0
     ninja==1.11.1.1
     notebook==7.0.6
     notebook-shim==0.2.3
     numba==0.57.1
     numpy==1.24.3
     omegaconf==2.3.0
     opencv-python==4.8.0.76
     pandas==2.0.3
     pybullet-svl==3.1.6.4
     pynput==1.7.6
     pytorch-lightning==2.1.0
     pyyaml==6.0.1
     scikit-learn==1.3.0
     scipy==1.10.1
     tensorboard==2.13.0
     tensorboard-data-server==0.7.2
     tensorflow==2.13.1
     tensorflow-estimator==2.13.0
     tensorflow-hub==0.16.1
     tensorflow-io-gcs-filesystem==0.34.0
     tensorflow-metadata==1.14.0
     termcolor==2.3.0
     tf-keras==2.15.0
     tfds-nightly==4.9.2.dev202308090034
     timm==0.9.6
     tokenizers==0.12.1
     torch==2.0.1
     torchvision==0.15.2
     tqdm==4.66.1
     transformers==4.31.0
     wandb==0.15.9